Migrate CAReduce to the C dispatch registry
CAReduce becomes a plain Op; NonZeroDimsCAReduce's _c_all override becomes an
error_on_empty_reduce_axis flag the generator reads. The elemwise and careduce
C generators move into link.c.dispatch.elemwise beside the impls, leaving
tensor/elemwise.py free of any C codegen.
